Remove useless cast.

+++ src/sinogram.cpp
@@ -40,7 +40,7 @@ Eigen::MatrixXf getSinogram(
         // Calculate and allocate the output matrix
         size_t a_steps = (size_t) std::floor(360 / a_stepsize);
         size_t p_steps = (size_t) std::floor(input.rows() / p_stepsize);
-        Eigen::MatrixXf output((int) p_steps, (int) a_steps);
+        Eigen::MatrixXf output(p_steps, a_steps);
 
         // Process all angles
         for (size_t a_step = 0; a_step < a_steps; a_step++) {
